+++ CHANGELOG.md
@@ -21,6 +21,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2023-03-26
 The inital changes made after creating the initial branch.
 
-### Changed
+### Added:
+- Installed Typescript
+- Implemented Typescript configuration
+  
+### Changed:
 - Updated the README.md
 - Updated Changelog
+- Updated all files for Typescript support
+
+### Installed Packages:
+#### Typescript:
+@babel/preset-typescript: v7.21.0
+Typescript: v5.0.2
+@types/node: v18.15.10
+@types/react: v18.0.30
+@types/react-dom: v18.0.11 
+@types/jest: v29.5.0
+ts-loader: v9.4.2